moved detect_machine_name() to clusters.py
